Changed token-related classes location to separate package.
